Get nav icons in position using Flexbox
Instead of the janky positioning that was being used before, these changes redesign the HTML and CSS to use Flexbox to get them inline instead.  This requires less extra CSS strucuturing, and ultimately make it easier to get the SVG icons aligned correctly.

One of the things I'm pretty proud of is the fact that the scaling of the SVGs is all done using Sass variables.  Instead of just hard-coding pixel amounts, I calculated the right look using pixels and then, once they were done, abstracted them out to variables that computer the right sizes as a proportion to the height of the icons.  This, along with relying on Flexbox for positioning and spacing, allows me to adjust just the height variable to scale and position everything correctly.  Yay for Sass variables!

What will be interesting to see is how well that setup works for scaling the icons smaller on mobiles sizes; maybe the size of the SVGs should be determined by the size of the container, which can be adjusted using responsive breakpoints?  We shall see...